Safari-609.1.11.1/LayoutTests/http/tests/security/window-name-after-same-origin-main-frame-navigation.html

<!DOCTYPE html>
<html>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }

        console.log("window.name should equal \"test\" after a same-origin main frame navigation.");
        document.location = "resources/window-name-test.html?127.0.0.1";
    </script>
</html>
