Safari-611.1.10/LayoutTests/http/tests/security/window-name-after-same-origin-main-frame-navigation-expected.txt

CONSOLE MESSAGE: window.name should equal "test" after a same-origin main frame navigation.
CONSOLE MESSAGE: window.name equals "test".
